Undo last bugfix, make a new one
What I thought fixed the issue I saw with piling metascopes looks to
have just prevented multi-line non-function tokens from being scoped,
and also probably just shifted the scope stack to hide the problem.

This isn't fully fixed.  If you define a token to a parened statement,
it still thinks the token is a function name and the statement is the
function params, but at least now it should stop that from making a
metascope that doesn't get closed.