Function defs can't be inside function defs
This is more targetted at function-like macros, but it's true in both
cases.

Made a version of the prototype context for function body to include
that doesn't reinclude the function context itself.  The list of
contexts in this new context can be further slimmed down, and should be
before I merge this into master.